Support P5 v2.x.x for the coming v5.x.x release of the wrapper
