+++ package.json
@@ -20,18 +20,18 @@
   },
   "scripts": {
     "build": "rimraf dist && pnpm build:component && pnpm build:demo",
-    "build:component": "tsc --noEmit && vite build --config config/vite/vite.component.config.ts",
-    "build:demo": "vite build --config config/vite/vite.demo.config.ts",
-    "ci": "pnpm format:check && pnpm lint && pnpm test && pnpm build",
-    "dev": "vite --config config/vite/vite.demo.config.ts --host",
+    "build:component": "tsc --noEmit && vite build --config config/vite/vite.config.ts --mode lib",
+    "build:demo": "vite build --config config/vite/vite.config.ts --mode demo",
+    "dev": "vite --config config/vite/vite.config.ts --host",
     "format": "pnpm prettier --write .",
     "format:check": "pnpm prettier --check .",
+    "integrate": "pnpm format:check && pnpm lint && pnpm test && pnpm build",
     "lint": "eslint --config config/eslint/eslint.config.ts",
     "lint:fix": "pnpm lint --fix",
     "prettier": "prettier --config config/prettier/prettier.json --ignore-path .gitignore",
-    "test": "vitest run --silent --config config/vite/vite.component.config.ts",
-    "test:coverage": "vitest run --silent --coverage --config config/vite/vite.component.config.ts",
-    "test:watch": "vitest watch --config config/vite/vite.component.config.ts"
+    "test": "vitest run --silent --config config/vite/vite.config.ts",
+    "test:coverage": "vitest run --silent --coverage --config config/vite/vite.config.ts",
+    "test:watch": "vitest watch --config config/vite/vite.config.ts"
   },
   "keywords": [
     "react",
@@ -82,7 +82,7 @@
     "react-error-boundary": "^6.0.0"
   },
   "peerDependencies": {
-    "p5": ">= 1.11.3",
+    "p5": ">= 2.0.0",
     "react": ">= 19.0.0",
     "react-dom": ">= 19.0.0"
   },
@@ -90,15 +90,14 @@
     "@babel/eslint-plugin": "^7.27.1",
     "@eslint/compat": "^1.3.2",
     "@eslint/js": "^9.33.0",
-    "@testing-library/jest-dom": "6.6.4",
+    "@testing-library/jest-dom": "6.7.0",
     "@testing-library/react": "^16.3.0",
     "@trivago/prettier-plugin-sort-imports": "^5.2.2",
     "@types/jest": "^30.0.0",
-    "@types/node": "^24.2.1",
-    "@types/p5": "^1.7.6",
-    "@types/react": "^19.1.9",
+    "@types/node": "^24.3.0",
+    "@types/react": "^19.1.10",
     "@types/react-dom": "^19.1.7",
-    "@typescript-eslint/eslint-plugin": "^8.39.0",
+    "@typescript-eslint/eslint-plugin": "^8.39.1",
     "@vitejs/plugin-react": "^5.0.0",
     "@vitest/coverage-v8": "^3.2.4",
     "babel-plugin-react-compiler": "19.1.0-rc.2",
@@ -107,8 +106,6 @@
     "eslint-plugin-react-compiler": "19.1.0-rc.2",
     "eslint-plugin-react-hooks": "^5.2.0",
     "gh-pages": "^6.3.0",
-    "jiti": "^2.5.1",
-    "js": "^0.1.0",
     "jsdom": "^26.1.0",
     "p5.capture": "^1.6.0",
     "p5.record.js": "^0.2.0",
@@ -117,8 +114,8 @@
     "react-dom": "19.1.1",
     "rimraf": "^6.0.1",
     "typescript": "^5.9.2",
-    "typescript-eslint": "^8.39.0",
-    "vite": "^7.1.1",
+    "typescript-eslint": "^8.39.1",
+    "vite": "^7.1.2",
     "vite-plugin-dts": "^4.5.4",
     "vitest": "^3.2.4",
     "vitest-canvas-mock": "^0.3.3"
