+++ src/contracts/InputProps.ts
@@ -1,8 +1,7 @@
+import { type Sketch } from "@contracts/Sketch";
+import { type SketchProps } from "@contracts/SketchProps";
 import { type ReactNode } from "react";
 import { type FallbackProps } from "react-error-boundary";
-
-import { type Sketch } from "./Sketch";
-import { type SketchProps } from "./SketchProps";
 
 export type InputProps<Props extends SketchProps> = Props & {
   sketch?: Sketch<Props>;
