+++ src/contracts/P5CanvasInstance.ts
@@ -1,7 +1,7 @@
-import type p5 from "p5";
-
-import { type SketchProps } from "./SketchProps";
+import p5 from "@contracts/p5";
+import { type SketchProps } from "@contracts/SketchProps";
 
 export type P5CanvasInstance<Props extends SketchProps = SketchProps> = p5 & {
   updateWithProps?: (props: Props) => void;
+  [key: string]: unknown;
 };
