+++ src/contracts/P5WrapperProps.ts
@@ -1,6 +1,6 @@
-import { type InputProps } from "./InputProps";
-import { type SketchProps } from "./SketchProps";
-import { type WithChildren } from "./WithChildren";
+import { type InputProps } from "@contracts/InputProps";
+import { type SketchProps } from "@contracts/SketchProps";
+import { type WithChildren } from "@contracts/WithChildren";
 
 export type P5WrapperProps<Props extends SketchProps = SketchProps> =
   WithChildren<InputProps<Props>>;
