+++ src/contracts/p5.ts
@@ -1,0 +1,5 @@
+// @see https://github.com/processing/p5.js/pull/7863
+// @ts-expect-error: P5 does not export types correctly right now, should be resolved soon by the PR above.
+import p5 from "p5";
+
+export default p5;
