+++ tests/utils/removeCanvasInstance.test.ts
@@ -1,10 +1,10 @@
-import p5 from "p5";
+import { type CanvasInstanceRef } from "@contracts/CanvasInstanceRef";
+import p5 from "@contracts/p5";
+import { removeCanvasInstance } from "@utils/removeCanvasInstance";
 import { createRef } from "react";
 import { describe, expect, it, vi } from "vitest";
 
-import { type CanvasInstanceRef } from "../../src/contracts/CanvasInstanceRef";
 import { SketchProps } from "../../src/main";
-import { removeCanvasInstance } from "../../src/utils/removeCanvasInstance";
 
 describe("removeCanvasInstance", () => {
   it("Calls the remove method on the canvas if it exists", () => {
